force flag added

+++ commands/rm/rm.go
@@ -1,6 +1,7 @@
 package rm
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 
@@ -8,7 +9,8 @@ import (
 )
 
 const (
-	CmdRm = "rm"
+	CmdRm                = "rm"
+	Unwritable_File_Perm = 0444
 )
 
 func Command() *cli.Command {
@@ -28,11 +30,19 @@ func Command() *cli.Command {
 			},
 			&cli.BoolFlag{
 				Name:        "recursive",
-				Aliases:     []string{"r"},
+				Aliases:     []string{"R"},
 				Usage:       "remove directories and their contents recursively",
 				Required:    false,
 				DefaultText: "false",
 			},
+			// TODO: implement forcefully recursive
+			// &cli.BoolFlag{
+			// 	Name:        "forcefully_recursive",
+			// 	Aliases:     []string{"rf"},
+			// 	Usage:       "remove directories and their contents forcefully recursively",
+			// 	Required:    false,
+			// 	DefaultText: "false",
+			// },
 		},
 		HideHelp:        true,
 		HideHelpCommand: true,
@@ -40,13 +50,16 @@ func Command() *cli.Command {
 }
 
 func Action(c *cli.Context) error {
-	if c.Bool("recursive") {
-		if err := removeDirectories(c.Args().Slice()); err != nil {
+	force := c.Bool("force") || c.Bool("forcefully_recursive")
+	recursive := c.Bool("recursive") || c.Bool("forcefully_recursive")
+
+	if recursive {
+		if err := removeDirectories(c.Args().Slice(), force); err != nil {
 			cli.ShowAppHelp(c)
 			return err
 		}
 	} else {
-		if err := removeFiles(c.Args().Slice()); err != nil {
+		if err := removeFiles(c.Args().Slice(), force); err != nil {
 			cli.ShowAppHelp(c)
 			return err
 		}
@@ -55,9 +68,9 @@ func Action(c *cli.Context) error {
 	return nil
 }
 
-func removeFiles(files []string) error {
+func removeFiles(files []string, force bool) error {
 	for _, file := range files {
-		if err := removeFile(file, false); err != nil {
+		if err := removeFile(file, false, force); err != nil {
 			return err
 		}
 	}
@@ -65,9 +78,9 @@ func removeFiles(files []string) error {
 	return nil
 }
 
-func removeDirectories(dirs []string) error {
+func removeDirectories(dirs []string, force bool) error {
 	for _, dir := range dirs {
-		if err := removeDirectoryWithContents(dir); err != nil {
+		if err := removeDirectoryWithContents(dir, force); err != nil {
 			return err
 		}
 	}
@@ -75,7 +88,7 @@ func removeDirectories(dirs []string) error {
 	return nil
 }
 
-func removeDirectoryWithContents(dir string) error {
+func removeDirectoryWithContents(dir string, force bool) error {
 	// Open the file and get the file info
 	file, err := os.Open(dir)
 	if err != nil {
@@ -90,16 +103,16 @@ func removeDirectoryWithContents(dir string) error {
 
 	// If the file is not a directory, remove it as a file
 	if !stat.IsDir() {
-		return removeFile(dir, false)
+		return removeFile(dir, false, force)
 	}
 
 	// Get all files in the directory
-	names, err := file.Readdirnames(-1)
+	fileNames, err := file.Readdirnames(-1)
 	if err != nil {
 		return err
 	}
 
-	for _, name := range names {
+	for _, name := range fileNames {
 		stat, err := os.Stat(filepath.Join(dir, name))
 		if err != nil {
 			return err
@@ -107,25 +120,37 @@ func removeDirectoryWithContents(dir string) error {
 
 		// If the file is a directory, remove it recursively
 		if stat.IsDir() {
-			if err := removeDirectoryWithContents(filepath.Join(dir, name)); err != nil {
+			if err := removeDirectoryWithContents(filepath.Join(dir, name), force); err != nil {
 				return err
 			}
-		}
-
-		if err := removeFile(filepath.Join(dir, name), true); err != nil {
-			return err
 		}
 	}
 
 	// Remove top directory
-	if err := removeFile(dir, true); err != nil {
+	if err := removeFile(dir, true, force); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func removeFile(file string, isDir bool) error {
+func removeFile(file string, isDir bool, force bool) error {
+	if !force {
+		stat, err := os.Stat(file)
+		if err != nil {
+			return err
+		}
+
+		// if unwritable, ask to user if he wants to force remove
+		if stat.Mode().Perm() == Unwritable_File_Perm {
+			var answer string
+			print("rm: remove write-protected regular file '" + file + "'?(y/n) ")
+			if fmt.Scan(&answer); answer != "y" {
+				return nil
+			}
+		}
+	}
+
 	if isDir {
 		return os.RemoveAll(file)
 	}
